@@ -1,6 +1,6 @@
 [tool.black]
 line-length = 100
-target-version = ["py310"]
+target-version = ["py311"]
 exclude='\.eggs|\.git|\.mypy_cache|\.tox|\.venv|_build|buck-out|build|dist'
 
 [tool.isort]
