@@ -34,10 +34,10 @@ setup(
         "License :: OSI Approved :: Apache Software License",
         "Operating System :: OS Independent",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
         "Programming Language :: Python",
     ],
-    python_requires=">=3.10",
+    python_requires=">=3.11",
     packages=find_packages(),
     include_package_data=True,
     install_requires=[
